Update 1.0.4
A big GoFlow update that is mostly just reorganitation and clean up from the slowly worsening spaghetti code. Broke up the main code into several small modules that just get called. Also converted the entire architecture to be a http based server that is able to process api requests that will be needed in the future.